moved files to new directory and removed unused file
