+++ pkg/solana/chainreader/account_read_binding.go
@@ -1,4 +1,4 @@
-package solana
+package chainreader
 
 import (
 	"context"
