+++ pkg/solana/chainreader/bindings.go
@@ -1,4 +1,4 @@
-package solana
+package chainreader
 
 import (
 	"context"
