+++ pkg/solana/chainreader/bindings_test.go
@@ -1,4 +1,4 @@
-package solana
+package chainreader
 
 import (
 	"context"
