+++ pkg/solana/chainreader/chain_reader.go
@@ -1,4 +1,4 @@
-package solana
+package chainreader
 
 import (
 	"context"
